first pop up
wundert euch nicht, dass der canvas in der scene so riesig ist. hat wohl was mit der camera zu tun, is mir jetzt wurscht.
