+++ Assets/Scenes/Escape Room.unity
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:721eee7eddc2ab0719db4d1058e861e6273e0f1cde8344b3c33ab2f15e29603e
-size 430364
+oid sha256:f1c43521c6419ca2c176ec2062f5be0926ae383683e0470ee035f2452ff79b00
+size 440131
