+++ Assets/TextMesh Pro/Resources/Fonts & Materials/LiberationSans SDF - Fallback.asset
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:6f33cdcd6d203e175a012087a4209442c2b6811baeebce6c875bb8c8d7c64e51
-size 9266
+oid sha256:0315a915b2278560e4710bf5c3595966fe744adaee498b582128c77b11705081
+size 559841
